int64 -> int32 for ids and merge_pairs

+++ minbpe.py
@@ -99,8 +99,8 @@ class BasicTokenizer:
         text_bytes = text.encode("utf-8") # raw bytes
         ids = list(text_bytes) # list of integers in range 0..255
 
-        ids = torch.tensor(ids, dtype=torch.int64, device=device)
-        merge_pairs = torch.zeros((num_merges, 2), dtype=torch.int64, device=device)
+        ids = torch.tensor(ids, dtype=torch.int32, device=device)
+        merge_pairs = torch.zeros((num_merges, 2), dtype=torch.int32, device=device)
         false_tensor = torch.tensor([False], dtype=torch.bool, device=device)
 
         for i in range(num_merges):
